# cedar-submission-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe submission server takes CEDAR metadata to the repositories that expect it in their own\nshape: NCBI BioSample and AIRR, ImmPort, and LINCS.\n\nTwo things happen here. Validation converts a CEDAR instance into the target format and returns\nthat repository's own verdict on it, sending nothing. Submission receives a package of files a\nchunk at a time, assembles it, and forwards it once the last chunk lands — so a successful\nresponse means the chunk was stored and any submission it completed was started, not that the\nrepository has accepted anything.\n",
    "title" : "CEDAR Submission Server API",
    "version" : "2.9.4"
  },
  "paths" : {
...
    "/command/upload-cairr-to-cedar" : {
      "post" : {
        "description" : "Receive the files of an AIRR submission, assemble them in CEDAR, and once they are all present send them to NCBI over FTP. The package arrives a chunk at a time, so a caller sends this repeatedly; the submission starts by itself once the last chunk lands. A 200 therefore means the chunk was stored and any submission it completed was started, not that the repository has accepted anything. A package must carry exactly one metadata file. Submissions are sent one at a time, through a queue.",
        "operationId" : "uploadCAIRRToCEDAR",
        "responses" : {
          "200" : {
            "description" : "The chunk was stored; any submission it completed has been started"
          },
          "400" : {
            "description" : "Not a multipart upload, or the package does not carry exactly one metadata file"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "The chunk could not be stored or the submission could not be prepared"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Upload an AIRR submission package and send it to NCBI",
        "tags" : [ "NCBI AIRR" ]
      }
    },
...
    "/command/validate-cairr" : {
      "post" : {
        "description" : "Convert a CEDAR AIRR instance into the NCBI submission XML and validate it, returning the report. Validation only: nothing is sent to NCBI here.",
        "operationId" : "validateCairr",
        "responses" : {
          "200" : {
            "description" : "The validation report"
          },
          "400" : {
            "description" : "The instance could not be read as an AIRR submission"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "The submission XML could not be generated"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Validate an AIRR instance for NCBI",
        "tags" : [ "NCBI AIRR" ]
      }
    },
    "/command/validate-lincs" : {
      "post" : {
        "description" : "Forward a CEDAR instance to the LINCS dataset validator and return what it says. The status is the validator's own, so its refusal is reported as it reported it, and CEDAR adds no judgement of its own.",
        "operationId" : "validateInstance",
        "responses" : {
          "200" : {
            "description" : "The LINCS validation report"
          },
          "400" : {
            "description" : "The LINCS validator rejected the instance"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "The LINCS validator could not be reached"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Validate an instance against the LINCS validator",
        "tags" : [ "LINCS" ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}